Assignment2/requirements.txt
flask
mysql-connector-python
sentence-transformers
faiss-cpu
chromadb
openai
transformers
torch
